completely ignore SIGINT
This way, Ctrl+C will send the SIGINT to both sh-log and the child process, and the child process will die or do graceful shutdown. When we forwarded the SIGINT to the child process, it would possibly receive two SIGINTs, resulting in ungraceful shutdown for some apps (namely all my neural net training programs).

Interestingly, if you kill sh-log with the `kill` command, the child process not only receives a SIGHUP but also possibly a SIGPIPE (since the stdout/stderr is dead).