server "wait" system
This is the first step towards implementing a streaming websocket API.

There were really two options for this:

 1. Streaming clients register event channels that receive all log messages.
 2. Streaming clients sleep until there's new messages in the log, and manually read and send new messages from the log.

In approach 1, streaming clients would definitely get all messages from all logging clients. However, there would be a problem. If a streaming client had a really slow connection, there would be "blockage"--the `sh-log` client would have to wait for the streaming client to read messages out of the queue before more messages could be written.

In approach 2, there is a chance that a slow streaming client might "miss" messages if messages are coming in faster than it can take them off the log. The advantage is that logging will never block to wait for a streaming client to receive the message.