fix annoying sh-log Ctrl+C catching
When you kill a Python process, it tends to print a bunch of crap to the console, especially if it was running a lot of threads. This can make sh-log take a long time to finish up, and you have to manually kill the sh-log process if you want it to stop early.

Now, pressing Ctrl+C a second time will force kill sh-log.